Fix memory leak in the timer
We were creating two futures, but only awaiting until the one of them
finished and not canceling the second.

Because `reset()` was never called, we were just creating more and
more futures that would never complete.

This change fixes this by canceling the second task immediately after
the first one is complete.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>