Add `take_while` and `drop_while` to `Receiver`
The `take_while()` method is just an alias for `filter()`, with the
intention to eliminate the ambiguity and make it more readable.

On the other hand, `drop_while()` is the negation of `take_while()` and
provided as a convenience and also to improve readability.

These names are widely popular and used in other programming languages
as well as the Python `itertools` module.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>